requirements.txt
phidata
python-dotenv
duckduckgo-search
beautifulsoup4
fastapi
uvicorn
groq
packaging
yfinance
openai